+++ scripts/test.sh
@@ -1,1 +1,7 @@
-# TODO: add linux test scripts
+#!/bin/bash
+
+export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:"`pwd`/linux"
+
+dart pub global run coverage:test_with_coverage
+
+dart pub global run coverage:format_coverage --packages=.dart_tool/package_config.json --lcov -i coverage/coverage.json -o coverage/lcov.info