+++ test/test.dart
@@ -1,0 +1,5 @@
+import 'package:opencv_dart/opencv_dart.dart' as cv;
+
+void main() {
+  final img = cv.imread("assets/1.jpg");
+}