+++ test/video_test.dart
@@ -6,20 +6,27 @@ void main() async{
   test('cv.VideoWriter.empty', () {
     final writer = cv.VideoWriter.empty();
     expect(writer.isOpened, equals(false));
+    writer.release();
   });
 
   test('cv.VideoWriter.open', () {
-    final writer = cv.VideoWriter.open("test/images/small.mp4", "MJPG", 60, (400, 300));
+    final writer = cv.VideoWriter.open("test/images/small.mp4", "mp4v", 60, (400, 300));
+    writer.release();
   });
 
   test('cv.VideoCapture.empty', () {
     final vc = cv.VideoCapture.empty();
     expect(vc.isOpened, false);
+    vc.release();
   });
 
   // test('cv.VideoCapture.create', () {
-  //   final vc = cv.VideoCapture.create(r"C:\DEV\flutter\opencv_dart\test\images\small.mp4");
+  //   final vc = cv.VideoCapture.empty();
+  //   vc.setProp(cv.CAP_PROP_FOURCC, vc.toCodec("mp4v"));
+  //   final success = vc.open("test/images/small.mp4");
+  //   expect(success, true);
   //   expect(vc.isOpened, true);
+  //   vc.release();
   // });
 
   // test('cv.VideoCapture.fromFile', () {