+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.0.4
+
+* adjust prebuilt binaries distribution, use `dart run opencv_dart:setup` to download prebuilt binaries
+* Finished: test for imgproc module
+* Finished: test for Linux
+
+## 0.0.3
+
+* more test for imgproc module
+
 ## 0.0.3-dev.0
 
 * Finished: test for core module